Move dependency on d3 out of `util.ts` so that it's only used on pages that need it
